Split computed macros into a separate module
